TEIIDTOOLS-712 removing the profile concept
and updating the swagger link to openapi
