+++ server/komodo-core/src/main/java/org/komodo/relational/profile/internal/SqlCompositionImpl.java
@@ -22,9 +22,9 @@ import org.komodo.core.internal.repository.Repository;
 import org.komodo.core.repository.KomodoObject;
 import org.komodo.core.repository.ObjectImpl;
 import org.komodo.core.repository.PropertyValueType;
+import org.komodo.relational.dataservice.SqlComposition;
 import org.komodo.relational.internal.RelationalObjectImpl;
 import org.komodo.relational.internal.TypeResolver;
-import org.komodo.relational.profile.SqlComposition;
 import org.komodo.spi.KException;
 import org.komodo.spi.repository.KomodoType;
 import org.komodo.spi.repository.UnitOfWork;
