+++ server/komodo-core/src/test/java/org/komodo/relational/workspace/WorkspaceManagerTest.java
@@ -26,15 +26,23 @@ import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
 
 import java.io.File;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+import org.komodo.core.KomodoLexicon;
 import org.komodo.core.repository.KomodoObject;
 import org.komodo.core.repository.ObjectImpl;
 import org.komodo.relational.RelationalModelTest;
 import org.komodo.relational.dataservice.Dataservice;
+import org.komodo.relational.dataservice.StateCommand;
+import org.komodo.relational.dataservice.StateCommandAggregate;
 import org.komodo.relational.dataservice.internal.DataserviceImpl;
+import org.komodo.relational.internal.RelationalModelFactory;
+import org.komodo.relational.profile.internal.StateCommandAggregateImpl;
+import org.komodo.relational.profile.internal.ViewEditorStateImpl;
 import org.komodo.spi.StringConstants;
 import org.komodo.utils.FileUtils;
 
@@ -116,6 +124,60 @@ public final class WorkspaceManagerTest extends RelationalModelTest {
     	assertTrue(wsMgr.deleteSchema("a"));
     	
     	assertNull(wsMgr.findSchema("a"));
+    }
+    
+    @Test
+    public void shouldAddViewEditorState() throws Exception {
+        String name = "myNewView";
+        ViewEditorStateImpl viewEditorState = RelationalModelFactory.createViewEditorState(getTransaction(),
+                                                                                                                                               _repo, wsMgr, name);
+        assertEquals(name, viewEditorState.getName());
+
+        String undoId = "removeSourceCommand";
+        String srcName = "myFabSource";
+        Map<String, String> undoArgs = new HashMap<>();
+        undoArgs.put("srcName", srcName);
+
+        String redoId = "addSourceCommand";
+        String srcPath = "connection=conn1/schema=public/table=customer";
+        Map<String, String> redoArgs = new HashMap<>();
+        redoArgs.put("srcName", srcName);
+        redoArgs.put("srcPath", srcPath);
+
+        StateCommandAggregateImpl stateCmdAgg = viewEditorState.addCommand();
+        StateCommand undoCommand = stateCmdAgg.setUndo(undoId, undoArgs);
+        StateCommand redoCommand = stateCmdAgg.setRedo(redoId, redoArgs);
+
+        StateCommandAggregate[] commands = viewEditorState.getCommands();
+        assertEquals(1, commands.length);
+        assertEquals(stateCmdAgg, commands[0]);
+        assertEquals(KomodoLexicon.StateCommandAggregate.NAME_PREFIX + 0,
+                                     stateCmdAgg.getName());
+
+        assertEquals(undoId, undoCommand.getId());
+        assertEquals(undoArgs, undoCommand.getArguments());
+        assertEquals(redoId, redoCommand.getId());
+        assertEquals(redoArgs, redoCommand.getArguments());
+    }
+
+    @Test
+    public void shouldGetViewEditorStates() throws Exception {
+        final int numViews = 5;
+
+        for ( int i = 0; i < numViews; ++i ) {
+            String name = "myNewView" + i;
+            wsMgr.addViewEditorState( name);
+        }
+
+        assertThat( this.wsMgr.getViewEditorStates( ).length, is( numViews ) );
+    }
+
+    @Test
+    public void shouldRemoveView() throws Exception {
+        String name = "myNewView";
+        this.wsMgr.addViewEditorState( name);
+        this.wsMgr.removeViewEditorState( name );
+        assertThat( this.wsMgr.getViewEditorStates( ).length, is( 0 ) );
     }
 
 }
