+++ server/komodo-server/src/main/java/org/komodo/metadata/DeployStatus.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.komodo.relational;
+package org.komodo.metadata;
 
 import java.util.ArrayList;
 import java.util.List;
