+++ server/komodo-server/src/main/java/org/komodo/metadata/MetadataInstance.java
@@ -23,7 +23,6 @@ import java.util.List;
 import org.komodo.metadata.query.QSResult;
 import org.komodo.metadata.runtime.TeiidDataSource;
 import org.komodo.metadata.runtime.TeiidVdb;
-import org.komodo.relational.DeployStatus;
 import org.komodo.spi.KException;
 import org.komodo.spi.StringConstants;
 import org.teiid.adminapi.Admin;
