+++ server/komodo-server/src/main/java/org/komodo/metadata/MetadataNamespaces.java
