+++ server/komodo-server/src/main/java/org/komodo/metadata/query/QSResult.java
