--- server/komodo-server/src/main/java/org/komodo/rest/CallbackTimeoutException.java
@@ -1,34 +1,0 @@
-/*
- * Copyright Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags and
- * the COPYRIGHT.txt file distributed with this work.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.komodo.rest;
-
-public class CallbackTimeoutException extends KomodoRestException {
-
-    /**
-     * 
-     */
-    private static final long serialVersionUID = 2005200485166253985L;
-
-    /**
-     * @param message
-     *        the error message (can be empty)
-     */
-    public CallbackTimeoutException() {
-        super("A transaction callback timeout occurred");
-    }
-}
