+++ server/komodo-server/src/main/java/org/komodo/rest/relational/response/vieweditorstate/RestSqlProjectedColumn.java
@@ -19,7 +19,7 @@ package org.komodo.rest.relational.response.vieweditorstate;
 
 import java.net.URI;
 
-import org.komodo.relational.profile.SqlProjectedColumn;
+import org.komodo.relational.dataservice.SqlProjectedColumn;
 import org.komodo.rest.RestBasicEntity;
 import org.komodo.spi.KException;
